add start hike button to trail screen
